+++ src/nupic/algorithms/SpatialPooler.hpp
@@ -86,9 +86,8 @@ namespace nupic
                         Real synPermActiveInc=0.05,
                         Real synPermConnected=0.1,
                         Real minPctOverlapDutyCycles=0.001,
-                        Real minPctActiveDutyCycles=0.001,
                         UInt dutyCyclePeriod=1000,
-                        Real maxBoost=10.0,
+                        Real boostStrength=0.0,
                         Int seed=1,
                         UInt spVerbosity=0,
                         bool wrapAround=true);
@@ -193,26 +192,14 @@ namespace nupic
                 its previously learned inputs are no longer ever active, or when
                 the vast majority of them have been "hijacked" by other columns.
 
-          @param minPctActiveDutyCycles A number between 0 and 1.0, used to set
-                a floor on how often a column should be activate. Periodically,
-                each column looks at the activity duty cycle of all other
-                columns within its inhibition radius and sets its own internal
-                minimal acceptable duty cycle to:
-
-                    minPctDutyCycleAfterInh * max(other columns' duty cycles).
-
-                On each iteration, any column whose duty cycle after inhibition
-                falls below this computed value will get its internal boost
-                factor increased.
-
           @param dutyCyclePeriod The period used to calculate duty cycles.
                 Higher values make it take longer to respond to changes in
                 boost. Shorter values make it potentially more unstable and
                 likely to oscillate.
 
-          @param maxBoost A number greater or equal than 1.0, used to control
-          the strength of boosting. No boosting is applied if maxBoost=1.0.
-          The strength of boosting increases as a function of maxBoost.
+          @param boostStrength A number greater or equal than 0, used to 
+          control boosting strength. No boosting is applied if it is set to 0.
+          The strength of boosting increases as a function of boostStrength.
           Boosting encourages columns to have similar activeDutyCycles as their
           neighbors, which will lead to more efficient use of columns. However,
           too much boosting may also lead to instability of SP outputs.
@@ -241,9 +228,8 @@ namespace nupic
                                   Real synPermActiveInc=0.1,
                                   Real synPermConnected=0.1,
                                   Real minPctOverlapDutyCycles=0.001,
-                                  Real minPctActiveDutyCycles=0.001,
                                   UInt dutyCyclePeriod=1000,
-                                  Real maxBoost=10.0,
+                                  Real boostStrength=0.0,
                                   Int seed=1,
                                   UInt spVerbosity=0,
                                   bool wrapAround=true);
@@ -480,14 +466,15 @@ namespace nupic
 
           @returns real number of the maximum boost value.
           */
-          Real getMaxBoost() const;
+          Real getBoostStrength() const;
 
           /**
-          Sets the maximum boost value.
+          Sets the strength of boost.
 
-          @param maxBoost real number of maximum boost value, must be larger than 1.0
+          @param boostStrength real number of boosting strength, 
+          must be larger than 0.0
           */
-          void setMaxBoost(Real maxBoost);
+          void setBoostStrength(Real boostStrength);
 
           /**
           Returns the iteration number.
@@ -669,21 +656,6 @@ namespace nupic
           void setMinPctOverlapDutyCycles(Real minPctOverlapDutyCycles);
 
           /**
-          Returns the minimum tolerated activity duty cycle, given as percent of
-          neighbors' activity duty cycle.
-
-          @returns minPctOverlapDutyCycles real number of the minimum tolerated activity duty cycle.
-          */
-          Real getMinPctActiveDutyCycles() const;
-          /**
-          Sets the minimum tolerated activity duty cycle, given as percent of
-          neighbors' activity duty cycle.
-
-          @param minPctActiveDutyCycles real number of the minimum tolerated activity duty cycle.
-          */
-          void setMinPctActiveDutyCycles(Real minPctActiveDutyCycles);
-
-          /**
           Returns the boost factors for all columns. 'boostFactors' size must
           match the number of columns.
 
@@ -742,21 +714,6 @@ namespace nupic
           @param minOverlapDutyCycles real array of the minimum overlap duty cycles for all columns.
           */
           void setMinOverlapDutyCycles(Real minOverlapDutyCycles[]);
-
-          /**
-          Returns the minimum activity duty cycles for all columns.
-          '_minActiveDutyCycles' size must match the number of columns.
-
-          @param minActiveDutyCycles real array to store the minimum activity duty cycles for all columns.
-          */
-          void getMinActiveDutyCycles(Real minActiveDutyCycles[]) const;
-          /**
-          Sets the minimum activity duty cycles for all columns.
-          '_minActiveDutyCycles' size must match the number of columns.
-
-          @param minActiveDutyCycles real array of the minimum activity duty cycles for all columns.
-          */
-          void setMinActiveDutyCycles(Real minActiveDutyCycles[]);
 
           /**
           Returns the potential mapping for a given column. 'potential' size
@@ -1232,29 +1189,32 @@ namespace nupic
                                  UInt activeArray[]);
 
           /**
-              Update the boost factors for all columns. The boost factors are used to
-              increase the overlap of inactive columns to improve their chances of
-              becoming active. and hence encourage participation of more columns in the
-              learning process. This is a line defined as: y = mx + b boost =
-              (1-maxBoost)/minDuty * dutyCycle + maxFiringBoost. Intuitively this means
-              that columns that have been active enough have a boost factor of 1, meaning
-              their overlap is not boosted. Columns whose active duty cycle drops too much
-              below that of their neighbors are boosted depending on how infrequently they
-              have been active. The more infrequent, the more they are boosted. The exact
-              boost factor is linearly interpolated between the points (dutyCycle:0,
-              boost:maxFiringBoost) and (dutyCycle:minDuty, boost:1.0).
-              @verbatim
+            Update the boost factors for all columns. The boost factors are used to
+            increase the overlap of inactive columns to improve their chances of
+            becoming active, and hence encourage participation of more columns in the
+            learning process. The boosting function is a curve defined as:
+            boostFactors = exp[ - boostStrength * (dutyCycle - targetDensity)]
+            Intuitively this means that columns that have been active at the target
+            activation level have a boost factor of 1, meaning their overlap is not
+            boosted. Columns whose active duty cycle drops too much below that of their
+            neighbors are boosted depending on how infrequently they have been active.
+            Columns that has been active more than the target activation level have
+            a boost factor below 1, meaning their overlap is suppressed
 
-                      boostFactor
-                          ^
-              maxBoost _  |
-                          |\
-                          | \
-                    1  _  |  \ _ _ _ _ _ _ _
-                          |
-                          +--------------------> activeDutyCycle
-                             |
-                      minActiveDutyCycle
+            The boostFactor depends on the activeDutyCycle via an exponential function:
+
+                        boostFactor
+                        ^
+                        |
+                        |\
+                        | \
+                  1  _  |  \
+                        |    _
+                        |      _ _
+                        |          _ _ _ _
+                        +--------------------> activeDutyCycle
+                              |
+                        targetDensity
               @endverbatim
             */
           void updateBoostFactors_();
@@ -1324,7 +1284,7 @@ namespace nupic
           UInt stimulusThreshold_;
           UInt inhibitionRadius_;
           UInt dutyCyclePeriod_;
-          Real maxBoost_;
+          Real boostStrength_;
           UInt iterationNum_;
           UInt iterationLearnNum_;
           UInt spVerbosity_;
@@ -1346,7 +1306,6 @@ namespace nupic
           vector<Real> minActiveDutyCycles_;
 
           Real minPctOverlapDutyCycles_;
-          Real minPctActiveDutyCycles_;
 
           SparseMatrix<UInt,Real,Int,Real64> permanences_;
           SparseBinaryMatrix<UInt, UInt> potentialPools_;